オブジェクトIDを生成する処理を追加
