+++ ocaz/bin/add.py
@@ -1,8 +1,14 @@
 #!/usr/bin/env python3
 
 import logging
+import os
+import sys
 
 import click
+
+
+sys.path.append(os.path.join(os.path.dirname(__file__), "..", "lib"))
+from ocaz.object_util import get_object_info
 
 
 @click.command()
@@ -33,6 +39,13 @@ def main(
     logging.info("log_level = %s", log_level)
     logging.debug("url = %s", url)
     logging.debug("data_dir = %s", data_dir)
+
+    object_info = get_object_info(url)
+    object_id = object_info["object_id"]
+    logging.debug("object_info = %s", object_info)
+    logging.info("object_id = %s", object_id)
+
+    logging.info("done")
 
 
 if __name__ == "__main__":
